moves translator module
